handle update operation mutation
